packages/chatten_rag/pyproject.toml
[project]
name = "chatten-rag"
dynamic = ["version"]
description = "Data extraction and RAG model on Databricks platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chatten",
    "databricks-agents>=0.15.0",
    "databricks-langchain>=0.3.0",
    "databricks-sdk>=0.43.0",
    "databricks-vectorsearch>=0.47",
    "langchain-text-splitters>=0.3.6",
    "langgraph>=0.2.70",
    "mlflow-skinny[databricks,langchain]>=2.20.1",
    "pandas>=2.2.3",
    "pypdf>=5.2.0",
    "tenacity>=9.0.0",
    "transformers>=4.48.3",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
root = "../../"

[tool.uv.sources]
chatten = { workspace = true }

[dependency-groups]
dev = [
    "pyspark>=3.5.4",
]

[project.scripts]
loader = "chatten_rag.tasks.loader:Loader.entrypoint"
indexer = "chatten_rag.tasks.indexer:Indexer.entrypoint"
driver = "chatten_rag.tasks.driver:Driver.entrypoint"
